move reducedPopups to their own file
