+++ src/app/services/context/popups-context/index.ts
@@ -28,4 +28,33 @@ export interface ShowPopupsParameters<Data = {}> {
   dialogProps: PopupsProps<Data>;
 }
 
+export interface PopupAction {
+  type: string;
+  payload: Popup;
+}
+
+export interface PopupState {
+  popups: Popup[];
+}
+
+export const POPUPS = {
+  ADD_POPUPS: 'ADD_POPUPS',
+  DELETE_POPUPS: 'DELETE_POPUPS'
+}
+
+export interface Popup<Data = {}> {
+  component: React.ElementType<PopupsComponent<Data>>,
+  popupsProps: PopupsProps<Data>,
+  type: PopupsType,
+  id: string;
+  result?: any;
+  open?: boolean;
+  status?: PopupsStatus;
+  reference?: string;
+}
+
+export const initialState: PopupState = {
+  popups: []
+};
+
 export { PopupsProvider };
